requirements.txt
asgiref==3.2.10
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
idna==2.10
lxml==4.5.2
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
